back/requirements.txt
Django==3.2.8
djangorestframework==3.12.4
psycopg2-binary==2.9.1
gunicorn==20.1.0
django-heroku==0.3.1
dj-database-url==0.5.0
drf-yasg==1.20.0
Pillow==8.4.0